Sentiment Analysis/ModelFront/requirements.txt
streamlit>=1.8.1
tensorflow-cpu==2.8.0

